+++ src/main.rs
@@ -89,13 +89,19 @@ pub struct QueryOpt {
     connect_opts: ConnectOpts,
     /// Size of a single batch in rows. The content of the data source is written into the output
     /// parquet files in batches. This way the content does never need to be materialized completely
-    /// in memory at once.
+    /// in memory at once. If `--batch-size-mib` is not specified this value defaults to 65535. This
+    /// avoids issues with some ODBC drivers using 16Bit integers to represent batch sizes. If
+    /// `--batch-size-mib` is specified no other limit is applied by default. If both option are
+    /// specified the batch size is the largest possible which satisfies both constraints.
     #[structopt(long)]
     batch_size_row: Option<usize>,
     /// Limits the size of a single batch. It does so by calculating the amount of memory each row
     /// requires in the allocated buffers and then limits the maximum number of rows so that the
     /// maximum buffer size comes as close as possbile, but does not exceed the specified amount.
-    /// Default is 2 GiB on 64 Bit platforms and 1 GiB on 32 Bit Platforms.
+    /// Default is 2 GiB on 64 Bit platforms and 1 GiB on 32 Bit Platforms if `--batch-size-row` is
+    /// not specified. If `--batch-size-row` is not specified no memory limit is applied by default.
+    /// If both option are specified the batch size is the largest possible which satisfies both
+    /// constraints.
     #[structopt(long)]
     batch_size_mib: Option<u32>,
     /// Maximum number of batches in a single output parquet file. If this option is omitted or 0 a