+++ Changelog.md
@@ -2,7 +2,8 @@
 
 ## Next
 
-* Both `--batch-size-row` and `--batch-size-mib` can now be specified
+* Both `--batch-size-row` and `--batch-size-mib` can now be both specified together.
+* If no `--batch-size-*` limit is specified. A row limit of 65535 is now also applied by default, next to the size limit.
 
 ## 0.6.8